requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.7.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
av==11.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
distro==1.9.0
dnspython==2.5.0
etils==1.6.0
flatbuffers==23.5.26
flax==0.8.1
fonttools==4.48.1
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
jax==0.4.24
jaxlib==0.4.23
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.11.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.1.9
orbax-checkpoint==0.4.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.23.4
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
pylibsrtp==0.10.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
streamlit==1.30.0
streamlit-audiorec==0.1.3
streamlit-webrtc==0.47.1
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_decision_forests==1.8.1
tensorflowjs==4.17.0
tensorstore==0.1.45
termcolor==2.4.0
tf-keras==2.15.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
wurlitzer==3.0.3
zipp==3.17.0